# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Malamh — Facial Consent Registry API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /user/profile:
    patch:
      operationId: updateProfile
      tags: [user]
      summary: Update user profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProfileBody"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /user/delete:
    delete:
      operationId: deleteAccount
      tags: [user]
      summary: Delete account and all data
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /u/{username}:
    get:
      operationId: getPublicProfile
      tags: [user]
      summary: Get public profile (no PII)
      parameters:
        - in: path
          name: username
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PublicProfile"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ─── Stats (wow endpoints) ────────────────────────────────────────────────────
...
  /internal/match:
    post:
      operationId: matchFace
      tags: [faces]
      summary: Match face against registry (playground endpoint)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MatchFaceBody"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MatchResult"
  /internal/consent-check:
    post:
      operationId: consentCheck
      tags: [faces]
      summary: AI Studio consent check — logs result, returns tokenId for polling
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ConsentCheckBody"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConsentCheckResult"

  # ─── API Keys ─────────────────────────────────────────────────────────────────
...
  /v1/check-face:
    post:
      operationId: checkFace
      tags: [public]
      summary: Check face consent (requires Bearer API key)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CheckFaceBody"
      responses:
        "200":
          description: Consent decision (no PII returned)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CheckFaceResponse"
        "401":
          description: Invalid API key
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /v1/check-face-with-token:
    post:
      operationId: checkFaceWithToken
      tags: [public]
      summary: Validate consent token and match face
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CheckFaceWithTokenBody"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CheckFaceResponse"
        "401":
          description: Invalid API key
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /v1/request-consent:
    post:
      operationId: requestConsent
      tags: [public]
      summary: Request consent for a face
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RequestConsentBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RequestConsentResponse"
        "401":
          description: Invalid API key
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required: [success]
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required: [error]

    # ─── Auth ────────────────────────────────────────────────────────────────────
...
    User:
      type: object
      properties:
        id:
          type: string
        email:
          type: string
        name:
          type: string
          nullable: true
        username:
          type: string
          nullable: true
        image:
          type: string
          nullable: true
        notifyOnScan:
          type: boolean
        notifyOnConsent:
          type: boolean
        notifyOnApiCheck:
          type: boolean
        createdAt:
          type: string
          format: date-time
        subscription:
          $ref: "#/components/schemas/Subscription"
          nullable: true
      required: [id, email, notifyOnScan, notifyOnConsent, notifyOnApiCheck, createdAt]
    UpdateProfileBody:
      type: object
      properties:
        name:
          type: string
          nullable: true
        username:
          type: string
          nullable: true
        notifyOnScan:
          type: boolean
        notifyOnConsent:
          type: boolean
        notifyOnApiCheck:
          type: boolean
    PublicProfile:
      type: object
      properties:
        username:
          type: string
        totalFaces:
          type: number
        blockedCount:
          type: number
        openCount:
          type: number
        tokenRequiredCount:
          type: number
        memberSince:
          type: string
          format: date-time
      required: [username, totalFaces, blockedCount, openCount, tokenRequiredCount, memberSince]

    # ─── Stats ───────────────────────────────────────────────────────────────────
...
    MatchFaceBody:
      type: object
      properties:
        imageBase64:
          type: string
        requesterName:
          type: string
        purpose:
          type: string
      required: [imageBase64]
    MatchResult:
      type: object
      properties:
        status:
          type: string
          enum: [blocked, open, token_required, no_match]
        matchScore:
          type: number
          nullable: true
        authUrl:
          type: string
          nullable: true
        tokenId:
          type: string
          nullable: true
        queryEmbeddingSample:
          type: array
          items:
            type: number
          nullable: true
        matchedEmbeddingSample:
          type: array
          items:
            type: number
          nullable: true
      required: [status]
    ConsentCheckBody:
      type: object
      properties:
        imageBase64:
          type: string
        prompt:
          type: string
        requesterName:
          type: string
      required: [imageBase64]
    ConsentCheckResult:
      type: object
      properties:
        status:
          type: string
          enum: [blocked, open, token_required, no_match]
        matchScore:
          type: number
          nullable: true
        authUrl:
          type: string
          nullable: true
        tokenId:
          type: string
          nullable: true
        logId:
          type: string
          nullable: true
      required: [status]

    # ─── API Keys ─────────────────────────────────────────────────────────────────
...
    Subscription:
      type: object
      properties:
        id:
          type: string
        plan:
          type: string
          enum: [FREE, MONITOR, MONITOR_PRO]
        status:
          type: string
        expiresAt:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
      required: [id, plan, status, createdAt]
...
    CheckFaceBody:
      type: object
      properties:
        imageBase64:
          type: string
          description: Base64-encoded image
        requesterName:
          type: string
        purpose:
          type: string
      required: [imageBase64]
    CheckFaceResponse:
      type: object
      properties:
        status:
          type: string
          enum: [blocked, open, token_required, no_match]
        matchScore:
          type: number
          nullable: true
        authUrl:
          type: string
          nullable: true
          description: Consent approval URL (only for token_required)
      required: [status]
    CheckFaceWithTokenBody:
      type: object
      properties:
        imageBase64:
          type: string
        consentToken:
          type: string
      required: [imageBase64, consentToken]
    RequestConsentBody:
      type: object
      properties:
        faceId:
          type: string
        requesterName:
          type: string
        purpose:
          type: string
      required: [faceId, requesterName]
    RequestConsentResponse:
      type: object
      properties:
        token:
          type: string
        authUrl:
          type: string
        expiresAt:
          type: string
          format: date-time
      required: [token, authUrl, expiresAt]